Add stuck timeout for jitter threads
On some hardware, if the clock fails the jitterentropy stuck test, we
can spin on the cpu for long periods of time during initalization.  To
avoid that, if we wait more than X seconds (configurable via command
line option) for a thread to report ready, mark it as stuck, and move
on, failing init if no threads start in time.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>